// src/components/lib/types/AnimationTypes.ts
export type Animations =
  | "flipIn"
  | "slitInVertical"
  | "scale"
  | "rotate"
  | "vibrate"
  | "blink"
  | "heartBeat"
  | "bounce"
  | "zoomTop"
  | "zoomBottom"
  | "zoomLeft"
  | "zoomRight"
  | "scaleInHorizontalCenter"
  | "expandIn"
  | "contractIn"
  | "expandOut"
  | "contractOut"
  | "focusIn"
  | "focusInExpand"
  | "focusInContract"
  | "focusOut"
  | "focusOutExpand"
  | "focusOutContract";